requirements.txt
# Core Google ADK (Agent Development Kit) dependencies
google-genai==1.50.1
google-adk[a2a]==1.18.0

# LangGraph and LangChain dependencies for state machine workflow
# Note: langgraph<1.0 required for compatibility with google-adk 1.18.0
langgraph==0.6.11
langchain-core==1.0.5

# Environment variable management
python-dotenv==1.2.1

# Data validation and structured outputs
pydantic==2.12.4

# FastMCP for Model Context Protocol server and client
fastmcp>=2.0.0

# Nest asyncio for running async code in sync context
nest-asyncio>=1.5.0

# A2A server and HTTP client dependencies
uvicorn>=0.27.0  # ASGI server for running A2A server
requests>=2.31.0  # HTTP client for testing A2A communication
